requirements.txt
discord.py==2.4.0
atoma==0.0.17
anthropic==0.18.1
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.3
asyncio-throttle==1.0.2 